+++ static_precompiler/tests/test_coffeescript.py
@@ -1,4 +1,6 @@
 # coding: utf-8
+import os
+
 import pytest
 
 from static_precompiler import compilers, exceptions
@@ -11,13 +13,30 @@ def clean_javascript(js):
     )
 
 
-def test_compile_file():
+def test_compile_file(monkeypatch, tmpdir):
+    monkeypatch.setattr("static_precompiler.settings.ROOT", tmpdir.strpath)
+
     compiler = compilers.CoffeeScript()
 
-    assert (
-        clean_javascript(compiler.compile_file("scripts/test.coffee")) ==
-        """(function() {\n  console.log("Hello, World!");\n}).call(this);"""
-    )
+    assert clean_javascript(compiler.compile_file("scripts/test.coffee")) == "COMPILED/scripts/test.js"
+    assert os.path.exists(compiler.get_full_output_path("scripts/test.coffee"))
+    with open(compiler.get_full_output_path("scripts/test.coffee")) as compiled:
+        assert clean_javascript(compiled.read()) == """(function() {\n  console.log("Hello, World!");\n}).call(this);"""
+
+
+def test_sourcemap(monkeypatch, tmpdir):
+
+    monkeypatch.setattr("static_precompiler.settings.ROOT", tmpdir.strpath)
+
+    compiler = compilers.CoffeeScript(sourcemap_enabled=False)
+    compiler.compile_file("scripts/test.coffee")
+    full_output_path = compiler.get_full_output_path("scripts/test.coffee")
+    assert not os.path.exists(os.path.splitext(full_output_path)[0] + ".map")
+
+    compiler = compilers.CoffeeScript(sourcemap_enabled=True)
+    compiler.compile_file("scripts/test.coffee")
+    full_output_path = compiler.get_full_output_path("scripts/test.coffee")
+    assert os.path.exists(os.path.splitext(full_output_path)[0] + ".map")
 
 
 def test_compile_source():