+++ CHANGES.rst
@@ -2,6 +2,16 @@
 Changes
 =======
 
+Dev
+===
+
+- Add source maps support for SASS/SCSS
+- Add source maps support for LESS
+- Add source maps support for CoffeeScript
+- Add source maps support for Stylus
+- Add source maps support for Babel
+
+
 1.0.1
 =====